rename file and create modularity
